DROID-3590 move
